made the BackgroundLogic class work with the main view model. This way its easier to read and work with.
Now i need to implement the background worker to ensure the hooks event are delt with quickly.
